+++ app/models/issue.rb
@@ -21,18 +21,6 @@ class Issue < WorkPackage
 
   attr_protected :project_id, :author_id, :lft, :rgt
 
-  validates_presence_of :subject, :priority, :project, :type, :author, :status
-
-  validates_length_of :subject, :maximum => 255
-  validates_inclusion_of :done_ratio, :in => 0..100
-  validates_numericality_of :estimated_hours, :allow_nil => true
-
-  validate :validate_format_of_due_date
-  validate :validate_start_date_before_due_date
-  validate :validate_start_date_before_soonest_start_date
-  validate :validate_fixed_version_is_assignable
-  validate :validate_fixed_version_is_still_open
-  validate :validate_enabled_type
 
   scope :open, :conditions => ["#{IssueStatus.table_name}.is_closed = ?", false], :include => :status
 
@@ -178,73 +166,12 @@ class Issue < WorkPackage
     Setting.issue_done_ratio == 'issue_field'
   end
 
-  def validate_format_of_due_date
-    if self.due_date.nil? && @attributes['due_date'] && !@attributes['due_date'].empty?
-      errors.add :due_date, :not_a_date
-    end
-  end
-
-  def validate_start_date_before_due_date
-    if self.due_date and self.start_date and self.due_date < self.start_date
-      errors.add :due_date, :greater_than_start_date
-    end
-  end
-
-  def validate_start_date_before_soonest_start_date
-    if start_date && soonest_start && start_date < soonest_start
-      errors.add :start_date, :invalid
-    end
-  end
-
-  def validate_fixed_version_is_assignable
-    if fixed_version
-      errors.add :fixed_version_id, :inclusion unless assignable_versions.include?(fixed_version)
-    end
-  end
-
-  def validate_fixed_version_is_still_open
-    if fixed_version && assignable_versions.include?(fixed_version)
-      errors.add :base, I18n.t(:error_can_not_reopen_work_package_on_closed_version) if reopened? && fixed_version.closed?
-    end
-  end
-
-  def validate_enabled_type
-    # Checks that the issue can not be added/moved to a disabled type
-    if project && (type_id_changed? || project_id_changed?)
-      errors.add :type_id, :inclusion unless project.types.include?(type)
-    end
-  end
-
   # Set the done_ratio using the status if that setting is set.  This will keep the done_ratios
   # even if the user turns off the setting later
   def update_done_ratio_from_issue_status
     if Issue.use_status_for_done_ratio? && status && status.default_done_ratio
       self.done_ratio = status.default_done_ratio
     end
-  end
-
-  # Return true if the issue is being reopened
-  def reopened?
-    if !new_record? && status_id_changed?
-      status_was = IssueStatus.find_by_id(status_id_was)
-      status_new = IssueStatus.find_by_id(status_id)
-      if status_was && status_new && status_was.is_closed? && !status_new.is_closed?
-        return true
-      end
-    end
-    false
-  end
-
-  # Return true if the issue is being closed
-  def closing?
-    if !new_record? && status_id_changed?
-      status_was = IssueStatus.find_by_id(status_id_was)
-      status_new = IssueStatus.find_by_id(status_id)
-      if status_was && status_new && !status_was.is_closed? && status_new.is_closed?
-        return true
-      end
-    end
-    false
   end
 
   # Is the amount of work done less than it should for the due date